feat: server no need uds client now
rewrite server, now when start server, uds client won't start anymore

+++ src/server/filesystem/attr.rs
@@ -1,4 +1,5 @@
 use std::os::unix::fs::MetadataExt;
+use std::os::unix::fs::PermissionsExt;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use async_std::fs::Metadata;
@@ -32,8 +33,7 @@ pub fn metadata_to_file_attr(inode: Inode, metadata: Metadata) -> Result<FileAtt
         },
         atime: metadata.accessed()?,
         mtime: metadata.modified()?,
-        ctime: UNIX_EPOCH
-            + Duration::new(metadata.ctime() as u64, metadata.ctime_nsec() as u32),
+        ctime: UNIX_EPOCH + Duration::new(metadata.ctime() as u64, metadata.ctime_nsec() as u32),
         perm: if metadata.is_dir() {
             (metadata.permissions().mode() ^ libc::S_IFDIR) as u16
         } else {
@@ -43,6 +43,6 @@ pub fn metadata_to_file_attr(inode: Inode, metadata: Metadata) -> Result<FileAtt
         gid: metadata.gid(),
         rdev: metadata.rdev() as u32,
         flags: 0,
-        nlink: metadata.nlink(),
+        nlink: metadata.nlink() as u32,
     })
 }
